fix old style log.d.ts jsvalue handling. and move inputString
