+++ apps/inspect/src/app/samples/sample-tools/filters.ts
@@ -4,7 +4,7 @@ import { Scores1 } from "../../../@types/log";
 import { FilterError, ScoreLabel } from "../../../app/types";
 import { SampleSummary } from "../../../client/api/types";
 import { kScoreTypeBoolean } from "../../../constants";
-import { inputString } from "../../../utils/format";
+import { inputString } from "@tsmono/inspect-common/utils";
 import { EvalDescriptor, ScoreDescriptor } from "../descriptor/types";
 
 import { kSampleMetadataPrefix } from "./sample-filter/language";
