+++ packages/inspect-common/src/utils/index.ts
@@ -1,1 +1,2 @@
 export { expandEvents } from "./expandEvents";
+export { inputString } from "./inputString";
