+++ packages/inspect-common/src/utils/inputString.ts
@@ -1,0 +1,30 @@
+import type { ChatMessage } from "../types";
+
+/**
+ * Gets a string for a sample input.
+ */
+export const inputString = (input: string | Array<ChatMessage>): string[] => {
+  if (typeof input === "string") {
+    return [input];
+  } else {
+    return input.map((inp) => {
+      if (typeof inp === "string") {
+        return inp;
+      } else {
+        const content = inp.content;
+        if (typeof content === "string") {
+          return content;
+        } else {
+          const result = content.map((con) => {
+            if (con.type === "text") {
+              return con.text;
+            } else {
+              return "";
+            }
+          });
+          return result.join("\n");
+        }
+      }
+    });
+  }
+};
